+++ spotifywebapipython/spotifyclient.py
@@ -21,6 +21,7 @@ from zeroconf import Zeroconf
 
 # our package imports.
 from .oauthcli import AuthClient
+from .vibrant import Vibrant, Palette
 from .models import *
 from .models import UserProfile as UserProfileCurrentUser
 from .spotifyconnect import SpotifyConnectDirectoryTask, SpotifyConnectDeviceNotFound
@@ -7215,6 +7216,122 @@ class SpotifyClient:
         except Exception:
             
             return None
+
+
+    def GetImageVibrantColors(
+        self, 
+        imageSource:str=None, 
+        colorCount:int=64, 
+        colorQuality:int=5, 
+        ) -> ImageVibrantColors:
+        """
+        Extracts vibrant color palette RGB values from the specified image source.  
+        
+        Args:
+            imageSource (str):  
+                The image source to extract color palette information from.  If the prefix of the 
+                value is `http:` or `https:`, then the image is downloaded from the url.  
+                This can also point to a filename on the local file system.  
+                If null, the currently playing Spotify track image url is used.  
+                Example: `http://mydomain/image1.jpg`  
+                Example: `c:/image1.jpg`  
+            colorCount (int):  
+                The number of colors in the initial palette from which swatches will be generated.  
+                Default is 64.
+            colorQuality (int):  
+                Controls the processing time and quality of the palette generation.  
+                A lower value (e.g. 1) results in higher quality but takes more processing time, 
+                while a higher value (e.g. 5) is faster but may result in a lower-quality palette.   
+                Default is 5.
+                
+        Returns:
+            A `ImageVibrantColors` object that contains extracted color information.
+                
+        Raises:
+            SpotifyWebApiError: 
+                If the Spotify Web API request was for a non-authorization service 
+                and the response contains error information.
+            SpotifyApiError: 
+                If the method fails for any other reason.
+
+        <details>
+          <summary>Sample Code</summary>
+        ```python
+        .. include:: ../docs/include/samplecode/SpotifyClient/GetImageVibrantColors.py
+        ```
+        </details>
+        """
+        apiMethodName:str = 'GetImageVibrantColors'
+        apiMethodParms:SIMethodParmListContext = None
+        result:ImageVibrantColors = {}
+        
+        try:
+            
+            # trace.
+            apiMethodParms = _logsi.EnterMethodParmList(SILevel.Debug, apiMethodName)
+            apiMethodParms.AppendKeyValue("imageSource", imageSource)
+            apiMethodParms.AppendKeyValue("colorCount", colorCount)
+            apiMethodParms.AppendKeyValue("colorQuality", colorQuality)
+            _logsi.LogMethodParmList(SILevel.Verbose, "Get vibrant colors from an image url", apiMethodParms)
+                
+            # validations.
+            if (not isinstance(colorCount,int)):
+                colorCount = 64
+            if (not isinstance(colorQuality,int)):
+                colorQuality = 5
+
+            # if image source not specified, then return currently playing track image url value.
+            if (imageSource is None):
+                nowPlaying:PlayerPlayState = self.GetPlayerNowPlaying(additionalTypes='episode')
+                if nowPlaying is not None:
+                    if nowPlaying.CurrentlyPlayingType in ['track','episode']:
+                        trackItem:Track = nowPlaying.Item
+                        if (trackItem is not None):
+                            imageSource = trackItem.ImageUrl
+            if (imageSource is None):
+                raise SpotifyApiError(SAAppMessages.ARGUMENT_REQUIRED_ERROR % (apiMethodName, 'imageSource'), logsi=_logsi)
+
+            # was a string value specified?
+            if (isinstance(imageSource, str)):
+                
+                imageSourceTmp = imageSource.lower()
+                
+                # was a url specified? 
+                if (imageSourceTmp.startswith("http:") or imageSourceTmp.startswith("https:")):
+                    
+                    # get image data from the url.
+                    response:HTTPResponse = self._Manager.request("GET", imageSource)
+                    if (response.status != 200):
+                        raise SpotifyApiError("Image URL could not be accessed (%s - %s): \"%s\"" % (response.status, response.reason, imageSource), logsi=_logsi)
+                    imageSource = response.data
+
+            # prepare to extract vibrant colors from the image source.
+            vibrant = Vibrant(color_count=colorCount, quality=colorQuality)
+
+            # extract the color palette; this returns a Palette containing vibrant, muted, 
+            # dark vibrant, dark muted, light vibrant, and light muted colors.
+            palette:Palette = vibrant.get_palette(imageSource)
+            _logsi.LogObject(SILevel.Verbose, TRACE_METHOD_RESULT_TYPE % (apiMethodName, type(palette).__name__), palette, excludeNonPublic=True)
+        
+            # process results.
+            result = ImageVibrantColors(root=palette)
+
+            # trace.
+            _logsi.LogObject(SILevel.Verbose, TRACE_METHOD_RESULT_TYPE % (apiMethodName, type(result).__name__), result, excludeNonPublic=True)
+            return result
+
+        except SpotifyApiError: raise  # pass handled exceptions on thru
+        except SpotifyWebApiError: raise  # pass handled exceptions on thru
+        except SpotifyWebApiAuthenticationError: raise  # pass handled exceptions on thru
+        except Exception as ex:
+            
+            # format unhandled exception.
+            raise SpotifyApiError(SAAppMessages.UNHANDLED_EXCEPTION.format(apiMethodName, str(ex)), ex, logsi=_logsi)
+
+        finally:
+        
+            # trace.
+            _logsi.LeaveMethod(SILevel.Debug, apiMethodName)
 
 
     def GetMarkets(